Move all functions from main file to helpers
