+++ assets/src/amp-story-editor-blocks.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { every } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { addFilter } from '@wordpress/hooks';
@@ -12,7 +7,6 @@ import domReady from '@wordpress/dom-ready';
 import { select, subscribe, dispatch } from '@wordpress/data';
 import { registerPlugin } from '@wordpress/plugins';
 import {
-	createBlock,
 	getDefaultBlockName,
 	setDefaultBlockName,
 	getBlockTypes,
@@ -45,11 +39,13 @@ import {
 	addAMPExtraProps,
 	getTotalAnimationDuration,
 	renderStoryComponents,
-	getTagName,
+	maybeSetInitialPositioning,
+	maybeSetTagName,
+	maybeUpdateAutoAdvanceAfterMedia,
 	wrapBlocksInGridLayer,
 } from './helpers';
 
-import { ALLOWED_BLOCKS, ALLOWED_TOP_LEVEL_BLOCKS, ALLOWED_CHILD_BLOCKS, MEDIA_INNER_BLOCKS } from './constants';
+import { ALLOWED_BLOCKS } from './constants';
 
 import store from './stores/amp-story';
 
@@ -153,109 +149,6 @@ domReady( () => {
 
 	unregisterBlockStyle( 'core/quote', 'large' );
 } );
-
-const positionTopLimit = 75;
-const positionTopHighest = 0;
-const positionTopGap = 10;
-
-/**
- * Set initial positioning if the selected block is an unmodified block.
- *
- * @param {string} clientId Block ID.
- */
-function maybeSetInitialPositioning( clientId ) {
-	const block = getBlock( clientId );
-
-	if ( ! block || ! ALLOWED_CHILD_BLOCKS.includes( block.name ) ) {
-		return;
-	}
-
-	const parentBlock = getBlock( getBlockRootClientId( clientId ) );
-	// Short circuit if the top position is already set or the block has no parent.
-	if ( 0 !== block.attributes.positionTop || ! parentBlock ) {
-		return;
-	}
-
-	// Check if it's a new block.
-	const newBlock = createBlock( block.name );
-	const isUnmodified = every( newBlock.attributes, ( value, key ) => value === block.attributes[ key ] );
-
-	// Only set the position if the block was unmodified before.
-	if ( isUnmodified ) {
-		const highestPositionTop = parentBlock.innerBlocks
-			.map( ( childBlock ) => childBlock.attributes.positionTop )
-			.reduce( ( highestTop, positionTop ) => Math.max( highestTop, positionTop ), 0 );
-
-		// If it's more than the limit, set the new one.
-		const newPositionTop = highestPositionTop > positionTopLimit ? positionTopHighest : highestPositionTop + positionTopGap;
-
-		updateBlockAttributes( clientId, { positionTop: newPositionTop } );
-	}
-}
-
-/**
- * Verify and perhaps update autoAdvanceAfterMedia attribute for pages.
- *
- * For pages with autoAdvanceAfter set to 'media',
- * verify that the referenced media block still exists.
- * If not, find another media block to be used for the
- * autoAdvanceAfterMedia attribute.
- *
- * @param {string} clientId Block ID.
- */
-function maybeUpdateAutoAdvanceAfterMedia( clientId ) {
-	const block = getBlock( clientId );
-
-	if ( ! block || ! ALLOWED_TOP_LEVEL_BLOCKS.includes( block.name ) ) {
-		return;
-	}
-
-	if ( 'media' !== block.attributes.autoAdvanceAfter ) {
-		return;
-	}
-
-	const innerBlocks = getBlocksByClientId( getBlockOrder( clientId ) );
-
-	const mediaBlock = block.attributes.autoAdvanceAfterMedia && innerBlocks.find( ( { attributes } ) => attributes.anchor === block.attributes.autoAdvanceAfterMedia );
-
-	if ( mediaBlock ) {
-		return;
-	}
-
-	const firstMediaBlock = innerBlocks.find( ( { name } ) => MEDIA_INNER_BLOCKS.includes( name ) );
-	const autoAdvanceAfterMedia = firstMediaBlock ? firstMediaBlock.attributes.anchor : '';
-
-	if ( block.attributes.autoAdvanceAfterMedia !== autoAdvanceAfterMedia ) {
-		updateBlockAttributes( clientId, { autoAdvanceAfterMedia } );
-	}
-}
-
-/**
- * Determines the HTML tag name that should be used for text blocks.
- *
- * This is based on the block's attributes, as well as the surrounding context.
- *
- * For example, there can only be one <h1> tag on a page.
- * Also, font size takes precedence over text length as it's a stronger signal for semantic meaning.
- *
- * @param {string} clientId Block ID.
- */
-function maybeSetTagName( clientId ) {
-	const block = getBlock( clientId );
-
-	if ( ! block || 'amp/amp-story-text' !== block.name ) {
-		return;
-	}
-
-	const siblings = getBlocksByClientId( getBlockOrder( clientId ) ).filter( ( { clientId: blockId } ) => blockId !== clientId );
-	const canUseH1 = ! siblings.some( ( { attributes } ) => attributes.tagName === 'h1' );
-
-	const tagName = getTagName( block.attributes, canUseH1 );
-
-	if ( block.attributes.tagName !== tagName ) {
-		updateBlockAttributes( clientId, { tagName } );
-	}
-}
 
 let blockOrder = getBlockOrder();
 let allBlocksWithChildren = getClientIdsWithDescendants();
